Stop trying to get shutil.rmtree() to work on every platform/FS
I'm unable to reproduce the reported failure cases on the platforms I
have access to, even using the same downloaded subtitle files. I suspect
this is because differences not only on the platforms (Windows, Linux,
Mac OS X, Android) but also on filesystems and the (lack of) encodings
they use to represent paths.

Switch to a saner mechanism to create a safe and unique tempdir as
already provided by Python: tempfile.mkdtemp()

This means a new tempdir is created for every download attempt and, as
they aren't reused anymore thy will start to accumulate. Don't think
this will be a major problem, we'll see.